+++ backend/CHANGELOG.md
@@ -10,9 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Support for [Qdrant](https://qdrant.tech/) vector database.
-
 - Updated Milvus reference in the factory to be more robust to other Milvus datastores
-- 
+
+### Improved
+
+- Updating testing to Mock OpenAI client and embeddings
+
 ## [v0.1.4] - 2024-10-16
 
 ### Improved